+++ changelog.md
@@ -5,6 +5,7 @@
 * Added Alliance Ashenvale NPCs
 * Added Alliance Azshara NPCs
 * Added Theramore Isle, Dustwallow Marsh NPCs
+* Added Alliance Arathi Highlands NPCs
 * Added several missing NPCs to Stranglethorn Vale
 * Added several missing NPCs to Wetlands
 * Minor NPC fixes in Stormwind, Darnassus